refactor: moved formatSelectorCallstack to the common utils
