+++ src/errors/test-run/utils.js
@@ -27,29 +27,6 @@ export function formatUrl (url) {
     return `<a href="${url}">${url}</a>`;
 }
 
-export function formatSelectorCallstack (apiFnChain, apiFnIndex, viewportWidth) {
-    if (typeof apiFnIndex === 'undefined')
-        return '';
-
-    const emptySpaces    = 10;
-    const ellipsis       = '...)';
-    const availableWidth = viewportWidth - emptySpaces;
-
-    return apiFnChain.map((apiFn, index) => {
-        let formattedApiFn = String.fromCharCode(160);
-
-        formattedApiFn += index === apiFnIndex ? '>' : ' ';
-        formattedApiFn += ' | ';
-        formattedApiFn += index !== 0 ? '  ' : '';
-        formattedApiFn += apiFn;
-
-        if (formattedApiFn.length > availableWidth)
-            return formattedApiFn.substr(0, availableWidth - emptySpaces) + ellipsis;
-
-        return formattedApiFn;
-    }).join('\n');
-}
-
 export function formatExpressionMessage (expression, line, column) {
     const expressionStr = escapeHtml(expression);
